resolved all active comments
