+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import { Outlet, Route, Routes } from "react-router-dom";
 
 import { NavBar } from "./components/index.ts";
-import { Apply, Candidates, Home } from "./pages/index.ts";
+import { Apply, Candidates } from "./pages/index.ts";
 
 function Layout() {
   return (
@@ -17,8 +17,7 @@ function App() {
     <div className="app-container">
       <Routes>
         <Route path="/" element={<Layout />}>
-          <Route index element={<Home />} />
-          <Route path="home" element={<Home />} />
+          <Route index element={<Candidates />} />
           <Route path="apply" element={<Apply />} />
           <Route path="candidates" element={<Candidates />} />
         </Route>
