+++ frontend/src/components/InfoForm.tsx
@@ -1,0 +1,3 @@
+export function InfoForm() {
+  return <strong>Important Facts For Applicants BEFORE YOU APPLY:</strong>;
+}
