+++ frontend/src/pages/index.ts
@@ -1,5 +1,4 @@
 import { Apply } from "./Apply.tsx";
 import { Candidates } from "./Candidates.tsx";
-import { Home } from "./Home.tsx";
 
-export { Home, Candidates, Apply };
+export { Candidates, Apply };
